Replace ListGroup with BorderedTable
